Enhance the tokenizer so that it is able to deal with individual tokens larger than 2GiB.
FossilOrigin-Name: 4e713558f6c6ea3dfbc97d2eb0a328955ef6ad6f2ad05517efe9ed4a60181440

+++ ext/rtree/rtree.c
@@ -62,7 +62,7 @@
 #else
   #include "sqlite3.h"
 #endif
-int sqlite3GetToken(const unsigned char*,int*); /* In the SQLite core */
+i64 sqlite3GetToken(const unsigned char*,int*); /* In the SQLite core */
 
 #include <stddef.h>
